requirements.txt
python-dotenv
notion-client
pandas
transformers
nltk
scikit-learn
tqdm
torch
